improved project structure
