+++ __init__.py
@@ -1,0 +1,3 @@
+"""Terraform Cloud MCP - A MCP server for Terraform Cloud"""
+
+__version__ = '0.1.0'
