+++ api/__init__.py
@@ -1,0 +1,1 @@
+"""API client module for Terraform Cloud MCP"""
