+++ requirements.txt
@@ -1,3 +1,3 @@
-fastmcp
+mcp
 httpx>=0.24.0
 python-dotenv>=1.0.0
