+++ tools/__init__.py
@@ -1,0 +1,5 @@
+"""MCP tools for Terraform Cloud"""
+
+# Import tools for easier access
+import tools.auth as auth
+import tools.workspaces as workspaces
